Reduce memory footprint: float32 FFT + explicit heap trim after each poll
Two changes, benchmarked before/after against real DMI data (same
methodology as the disk-cache fix: cold poll, warm poll, and a
6-iteration repeated-cycle loop to separate a real plateau from a leak):

1. compute_motion (nowcast.py, both copies) now does its FFT-based phase
   correlation in float32/complex64 instead of numpy's float64/complex128
   default. This is peak-finding on a correlation surface, not
   precision-sensitive science, and it was the single biggest transient
   allocation in a poll cycle. Isolated to that stage: ~93% reduction in
   its own memory delta (+15.9MB -> +1.2MB).

2. gc.collect() + malloc_trim(0) at the end of every poll (coordinator.py
   and dev/server.py, same two authoritative points as the cache-pruning
   fix). Doesn't reduce the peak during active computation, but forces
   glibc to actually return freed heap arenas to the OS afterward instead
   of holding them in reserve for the ~2-minute idle gap until the next
   poll -- which is where the process spends nearly all its time.

Combined effect (repeated-cycle loop, same frame pair reused every
iteration): peak during active computation ~325MB -> ~318MB (modest,
expected -- most stages besides the FFT weren't touched); RSS at rest
between polls ~325MB (never released before) -> ~78MB, a ~76% reduction,
landing close to the ~60-76MB cold-process baseline.

One existing test needed loosening: float32's coarser precision made a
previously-strict `>` land exactly on the boundary value in
test_motion_field_blending_stays_directionally_distinct_around_the_global_
vector -- a precision artifact in the assertion, not a behavior
regression (the paired right-cell assertion already used <=). Full
before/after numbers and methodology in PLAN-HA-COMPONENT.md.

+++ dev/nowcast.py
@@ -30,8 +30,16 @@ def compute_motion(dbz_prev: np.ndarray, dbz_curr: np.ndarray,
     equally plausible), high at a sharp, distinctive edge — see
     estimate_motion_field's confidence-weighted blending, which is why
     this exists."""
-    a = dbz_prev.astype(np.float64)
-    b = dbz_curr.astype(np.float64)
+    # float32/complex64 rather than the numpy FFT default of float64/
+    # complex128: this is peak-finding on a correlation surface, not
+    # precision-sensitive science, and this array is the single biggest
+    # transient allocation in a poll cycle (see PLAN-HA-COMPONENT.md's
+    # resource-usage benchmark) — halving it directly halves that peak.
+    # numpy's fft2 preserves the input's precision family (float32 in ->
+    # complex64 out), so casting here is enough to keep everything below
+    # at half width.
+    a = dbz_prev.astype(np.float32)
+    b = dbz_curr.astype(np.float32)
     a = a - a.mean()
     b = b - b.mean()
 
@@ -41,7 +49,11 @@ def compute_motion(dbz_prev: np.ndarray, dbz_curr: np.ndarray,
     # correlation and pull the estimate toward zero shift even when the
     # actual rain clearly moved. Taper toward the array edges to suppress
     # that (standard fix for edge artifacts in phase correlation).
-    window = _hann2d(a.shape)
+    window = _hann2d(a.shape).astype(np.float32)  # np.hanning is float64 by
+                                                    # default; cast or the
+                                                    # multiply below upcasts
+                                                    # a/b straight back to
+                                                    # float64.
     a = a * window
     b = b * window
